+++ src/main/java/org/torproject/onionoo/docs/SummaryDocument.java
@@ -322,6 +322,18 @@ public class SummaryDocument extends Document {
     return this.hostName;
   }
 
+  @Expose
+  @SerializedName("rv")
+  private Boolean recommendedVersion;
+
+  public void setRecommendedVersion(Boolean recommendedVersion) {
+    this.recommendedVersion = recommendedVersion;
+  }
+
+  public Boolean getRecommendedVersion() {
+    return this.recommendedVersion;
+  }
+
   /* The familyFingerprints parameter can go away after September 8, 2015.
    * See above. */
   /** Instantiates a summary document with all given properties. */
@@ -330,7 +342,8 @@ public class SummaryDocument extends Document {
       boolean running, SortedSet<String> relayFlags, long consensusWeight,
       String countryCode, long firstSeenMillis, String asNumber,
       String contact, SortedSet<String> familyFingerprints,
-      SortedSet<String> effectiveFamily, String version, String hostName) {
+      SortedSet<String> effectiveFamily, String version, String hostName,
+      Boolean recommendedVersion) {
     this.setRelay(isRelay);
     this.setNickname(nickname);
     this.setFingerprint(fingerprint);
@@ -347,6 +360,7 @@ public class SummaryDocument extends Document {
     this.setEffectiveFamily(effectiveFamily);
     this.setVersion(version);
     this.setHostName(hostName);
+    this.setRecommendedVersion(recommendedVersion);
   }
 }