+++ src/test/java/org/torproject/onionoo/server/ResourceServletTest.java
@@ -144,7 +144,7 @@ public class ResourceServletTest {
             "0025C136C1F3A9EEFE2AE3F918F03BFA21B5070B" })),
         new TreeSet<>(Arrays.asList(
         new String[] { "001C13B3A55A71B977CA65EC85539D79C653A3FC" })),
-        "0.2.3.25", "ppp-62-216-201-221.dynamic.mnet-online.de");
+        "0.2.3.25", "ppp-62-216-201-221.dynamic.mnet-online.de", true);
     this.relays.put("000C5F55BD4814B917CC474BD537F1A3B33CCE2A",
         relayTorkaZ);
     org.torproject.onionoo.docs.SummaryDocument relayFerrari458 =
@@ -159,7 +159,7 @@ public class ResourceServletTest {
             "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })),
         new TreeSet<>(Arrays.asList(new String[] {
             "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })), null,
-            "c-68-38-171-200.hsd1.in.comcast.net");
+            "c-68-38-171-200.hsd1.in.comcast.net", null);
     this.relays.put("001C13B3A55A71B977CA65EC85539D79C653A3FC",
         relayFerrari458);
     org.torproject.onionoo.docs.SummaryDocument relayTimMayTribute =
@@ -172,7 +172,8 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-16 18:00:00"), "AS6830",
         "1024d/51e2a1c7 \"steven j. murdoch\" "
         + "<tor+steven.murdoch@cl.cam.ac.uk> <fb-token:5sr_k_zs2wm=>",
-        new TreeSet<String>(), new TreeSet<String>(), "0.2.3.24-rc-dev", null);
+        new TreeSet<String>(), new TreeSet<String>(), "0.2.3.24-rc-dev", null,
+        false);
     this.relays.put("0025C136C1F3A9EEFE2AE3F918F03BFA21B5070B",
         relayTimMayTribute);
     this.bridges = new TreeMap<>();
@@ -183,7 +184,7 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-21 18:07:03"), false,
         new TreeSet<>(Arrays.asList(new String[] { "Valid" })), -1L,
         null, DateTimeHelper.parse("2013-04-20 15:37:04"), null, null,
-        null, null, "0.2.2.39", null);
+        null, null, "0.2.2.39", null, true);
     this.bridges.put("0000831B236DFF73D409AD17B40E2A728A53994F",
         bridgeec2bridgercc7f31fe);
     org.torproject.onionoo.docs.SummaryDocument bridgeUnnamed =
@@ -193,7 +194,7 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-20 17:37:04"), false,
         new TreeSet<>(Arrays.asList(new String[] { "Valid" })), -1L,
         null, DateTimeHelper.parse("2013-04-14 07:07:05"), null, null,
-        null, null, null, null);
+        null, null, null, null, null);
     this.bridges.put("0002D9BDBBC230BD9C78FF502A16E0033EF87E0C",
         bridgeUnnamed);
     org.torproject.onionoo.docs.SummaryDocument bridgegummy =
@@ -204,7 +205,7 @@ public class ResourceServletTest {
         new TreeSet<>(Arrays.asList(new String[] { "Running",
             "Valid" })), -1L, null,
         DateTimeHelper.parse("2013-01-16 21:07:04"), null, null, null,
-        null, "0.2.4.4-alpha-dev", null);
+        null, "0.2.4.4-alpha-dev", null, false);
     this.bridges.put("1FEDE50ED8DBA1DD9F9165F78C8131E4A44AB756",
         bridgegummy);
   }
@@ -1694,6 +1695,29 @@ public class ResourceServletTest {
   @Test
   public void testHostNameUmlaut() {
     this.assertErrorStatusCode("/summary?host_name=äöü", 400);
+  }
+
+  @Test
+  public void testRecommendedVersionTrue() {
+    this.assertSummaryDocument("/summary?recommended_version=true", 1,
+        new String[] { "TorkaZ" }, 1, new String[] { "ec2bridgercc7f31fe" });
+  }
+
+  @Test
+  public void testRecommendedVersionFalse() {
+    this.assertSummaryDocument("/summary?recommended_version=false", 1,
+        new String[] { "TimMayTribute" }, 1, new String[] { "gummy" });
+  }
+
+  @Test
+  public void testRecommendedVersionTrueCapitalized() {
+    this.assertSummaryDocument("/summary?recommended_version=TRUE", 1,
+        new String[] { "TorkaZ" }, 1,new String[] { "ec2bridgercc7f31fe" });
+  }
+
+  @Test
+  public void testRecommendedVersionNull() {
+    this.assertErrorStatusCode("/summary?recommended_version=null", 400);
   }
 }